+++ noodles-bgzf/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+
+  * bgzf/reader: Verify the expected CRC32 checksum of the uncompressed data.
+
 ## 0.11.0 - 2022-03-29
 
 ### Added